Allow user overrides for ansible-role-requirements
Currently it is only possible to override the contents of the
ansible-role-requirements file as a whole.

This patch allows a user-role-requirements.yml file to be
located in the /etc/openstack_deploy directory (or the directory defined
by OSA_CONFIG_DIR). It is only necessary to list specific overridden
repos in the user role file. The boostrap-ansible.sh script is modified to
give preference to any repos named in user-role-requirements over those
in the standard ansible-role-requirements.

This gives the following benefits:

* When a deployer needs to use a locally patched version of an ansible
role, the configuration can be held under source code control in
/etc/openstack_deploy along with the rest of the environment
configuration

* Manual adjustments to ansible-role-requirements.yml are no longer lost
when running bootstrap-ansible.sh, the modifications can be kept in
userspace.

* If any additional ansible roles are required by the deployer above the
standard set these can be listed in user-role-requirements.yml

* It is very clear which roles have been overidden as they are the only
ones listed in user-role-requirements.yml

Change-Id: I689c7cf749f6611ae3dfcb1804a700d5e456947e